Remove Python introduction and basics
This commit removes a Python file that consisted of basic Python introduction and terminology. The file was in Polish language. It's been removed due to it being in a non-English language as well as covering basic content that already exists in the online resources widely. The goal is to keep the repository clean and maintain the content in English for wider accessibility.