Refactor Python classes and update data flow
Rearranged the code cells as an improvement in the educational flow. Execution cells have been cleared for a clean start for the students. Certain explanatory texts have been updated for better clarity and understanding, including renaming some of the objects in the example code for conformity with English language. These changes will significantly enhance the readability and maintainability of the code hence a better learning experience for the students.