Improve code execution visibility (outputs addition)
This commit increases the readability of the notebook and makes it easier to understand what is happening in each code segment. By adding outputs to each segment of code, we can visualize the expected results from running each block of code. This will serve as a valuable resource for users who may be learning these concepts for the first time, allowing them to match their understanding and expectations with tangible results.

This commit also fixes minor inconsistencies between comments and their corresponding codes. Such as the comment "Przeczytaj pierwsze n wierszy pliku * ", the "*" special character was misplaced and has been removed in this change.